requirements.txt
Flask==2.2.5
folium==0.14.0
fonttools==4.38.0
geographiclib==2.0
geopy==2.3.0
matplotlib==3.5.3
matplotlib-inline==0.1.6
musicbrainzngs==0.7.1
pandas==1.3.5
python-dateutil==2.8.2
python-dotenv==0.21.1
spotipy==2.23.0
waitress==2.1.2
Flask-Session==0.5.0